refactor: move token store to repositories
